Extract TurbineReceiveChannel and replace where appropriate
